Add ArgumentRegistrar Module
Module is responsible for all the human readable parts of the argument parser. Allows registration of arguments by name, tacking onto them either positions (in argc) or short/long switches (e.g. "-a" | "--all"). This will be managed alongside the ArgumentStore module such that keys in one are keys in the other, and registration to one will cause an entry in the other. In effect binding the "name" of an argument to a position or switch and a conversion function. The ArgumentParser to come will handle both of these objects and allow retrieval of all this information when given an arguments list.